+++ cmd/codeforces/list/submissions.go
@@ -3,6 +3,7 @@ package list
 import (
 	"fmt"
 	"os"
+	"strings"
 
 	"github.com/cp-tools/cpt-lib/v2/codeforces"
 	"github.com/cp-tools/cpt/util"
@@ -12,8 +13,8 @@ import (
 	"github.com/gosuri/uitable"
 )
 
-// SubmissionsMode displays tabular submission data.
-func SubmissionsMode(arg codeforces.Args, username string, count uint) {
+// Submissions displays tabular submission data.
+func Submissions(arg codeforces.Args, username string, count uint) {
 	// Anything more than 1 page (50 rows) makes no sense.
 	chanSubmissions, err := arg.GetSubmissions(username, 1)
 	if err != nil {
@@ -57,4 +58,25 @@ func SubmissionsMode(arg codeforces.Args, username string, count uint) {
 		fmt.Fprintln(writer, t.String())
 	}
 	writer.Stop()
+}
+
+// CompressVerdicts compresses and returns color to use.
+func CompressVerdicts(verdict string) string {
+	ccMap := [][]interface{}{
+		{"Accepted", "Accepted", color.FgHiGreen},
+		{"Partial result", "PR", color.FgHiGreen},
+		{"Compilation error", "CE", color.FgHiYellow},
+		{"Wrong answer", "WA", color.FgHiRed},
+		{"Runtime error", "RTE", color.FgHiRed},
+		{"Time limit exceeded", "TLE", color.FgHiYellow},
+	}
+
+	for _, val := range ccMap {
+		if key := val[0].(string); strings.Contains(verdict, key) {
+			// Clean text; update color; return
+			verdict = strings.ReplaceAll(verdict, key, val[1].(string))
+			return color.New(val[2].(color.Attribute)).Sprint(verdict)
+		}
+	}
+	return verdict
 }
